fix(ci): GRA-366 cargo fmt + ui-lint baseline + promote 8 Color32 literals to theme.rs (#227)
* fix(ci): cargo fmt + ui-lint baseline + promote 8 Color32 literals to theme.rs

Five operator direct-pushes (efa2140a..c6f2ebcc) on main left cargo fmt
red (56 diffs) and reintroduced 8 hardcoded Color32 literals outside
theme.rs. This commit restores CI-green state on c6f2ebcc:

1. `cargo fmt --all` — clears all 56 fmt diffs across 9 source files
   and 1 test file. Pure formatting, no behavior change.

2. audit_bevy_color_literals_baseline.txt — re-emits the baseline after
   the cascade shifted visuals.rs line numbers (the same
   line-keyed-baseline orphan pattern documented in
   `feedback-audit-baseline-orphan-check.md`). 11 trajectories-preview
   Color literals that already existed before the cascade are now
   in-baseline again.

3. ui/theme.rs — adds 3 new tokens for the 8 reintroduced literals:
     * `PORKCHOP_INFEASIBLE` (40, 40, 40) — dark-grey sentinel for
       porkchop cells with non-finite ΔV. Lifted from
       `porkchop_color_ramp.rs::INFEASIBLE_COLOR`; that name is now a
       re-export so the rest of the module keeps reading naturally.
     * `paused_overlay_fg(alpha)` — pause-button stroke colour, same
       hue as `RED` with caller-supplied alpha for the blink pulse.
     * `paused_label_color(alpha)` — slightly lighter red for the
       "⏸ PAUSED" label.
     * `color32_to_bevy(c)` — moved from porkchop_color_ramp.rs and
       re-exported there so callers don't construct their own
       `Color::srgba` literals.

4. ui/dashboard.rs — replaces the 2 hardcoded PAUSED literals
   (src/ui/dashboard.rs:1386, :1392) with the new theme helpers.

5. ui/porkchop_color_ramp.rs — replaces 3 procedural
   `Color32::from_rgba_unmultiplied` calls (lines 183, 200, 223 in
   the pre-cascade file) with `theme::lerp_rgba` /
   `theme::color32_from_rgba`. These were inside lambda interpolators
   that fed runtime components into the constructor, so the audit
   flagging them was over-strict; the theme funnel fixes both.

audit_color32_literals: 8 NEW → 0 NEW (all 8 promoted or removed).
audit_bevy_color_literals: 11 NEW → 0 NEW (baseline refreshed).

Verification:
- cargo fmt --all -- --check ✓
- python3 scripts/audit_color32_literals.py --strict ✓ (exit 0)
- python3 scripts/audit_bevy_color_literals.py --strict ✓ (exit 0)


* chore: retrigger CI for PR #227 (apt-get infra failure)

* fix(clippy): address second-order lint debt surfaced by CI

Three clippy errors flagged by `cargo clippy --all-targets --all-features
-- -D warnings` on c6f2ebcc:

1. `unused import: crate::ui::theme::color32_to_bevy` at
   src/ui/porkchop_color_ramp.rs:236 — the `pub use` re-export I added
   in the prior commit is unused (no external caller imports the symbol
   from this module). Removed the re-export; the function still lives at
   `crate::ui::theme::color32_to_bevy` with `#[allow(dead_code)]` so
   future test/diagnostic-dump callers can reach it.

2. `casting to the same type is unnecessary (u32 -> u32)` at
   src/persistence/snapshot.rs:164 — pre-existing latent clippy debt
   from the cascade. `bevy::ecs::entity::Entities::len()` already
   returns `u32`, so the `as u32` cast was redundant.

3. `manual_repeat_n` at src/ui/porkchop_color_ramp.rs:284 — clippy
   prefers `std::iter::repeat_n(8.0, 98)` over
   `std::iter::repeat(8.0).take(98)`. Trivial syntax swap.


* fix(clippy): address second-order lint debt surfaced by CI run 2

Two more clippy errors flagged on top of the previous fix:

1. `empty line after doc comment` at src/ui/porkchop_color_ramp.rs:234
   — the orphan doc comment I left when removing the `pub use
   color32_to_bevy` re-export. The `///` block was attached to
   `#[cfg(test)]` (a lint attribute, not a doc target). Removed the
   orphan comment.

2. `function cell_abs_t_dep is never used` at
   tests/porkchop_rotation_no_snap.rs:375 — pre-existing latent
   clippy debt from the cascade. Added `#[allow(dead_code)]` with a
   comment noting it's a helper for future cell-coordinate tests.


---------

+++ src/persistence/snapshot.rs
@@ -161,7 +161,7 @@ fn collect_all_entities(world: &World) -> Vec<bevy::prelude::Entity> {
     // is `Option<EntityLocation>` — `Some` means spawned.  Use the
     // public `get_spawned`-via-`resolve_from_index` round-trip
     // rather than poking at private fields.
-    for index in 0..entities.len() as u32 {
+    for index in 0..entities.len() {
         // `u32::MAX` is reserved as a "no entity" sentinel by
         // Bevy's entity allocator, so `EntityIndex::from_raw_u32`
         // would return `None`.  Skip it explicitly.
